# backend/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PA School Data API",
    "version": "1.0.0",
    "description": "Read-only JSON API behind chrispeterkins.com/paschools. Pennsylvania PSSA and Keystone results (2013/2015 to present), PVAAS growth, school directory data, and student-weighted aggregates. Responses are cacheable for an hour and carry ETags. Data: Pennsylvania Department of Education, NCES Common Core of Data, U.S. Census Bureau. See /paschools/about for methods and caveats."
  },
  "servers": [
    {
      "url": "https://chrispeterkins.com/paschools/api"
    }
  ],
...
  "components": {
    "parameters": {
      "exam": {
        "name": "exam",
        "in": "query",
        "schema": {
          "type": "string",
          "enum": [
            "pssa",
            "keystone"
          ],
          "default": "pssa"
        },
        "description": "PSSA (grades 3-8) or Keystone (grade 11)"
      },
...
      "subject": {
        "name": "subject",
        "in": "query",
        "schema": {
          "type": "string",
          "example": "Mathematics"
        },
        "description": "PSSA: Mathematics, English Language Arts, Science. Keystone: Algebra I, Biology, Literature."
      },
...
    },
...
  },
  "paths": {
...
    "/schools": {
      "get": {
        "tags": [
          "entities"
        ],
        "summary": "Search and page through schools",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "name",
                "districtName",
                "countyName",
                "type",
                "enrollment",
                "proficiency",
                "growth"
              ]
            }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "countyId",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "countyName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "districtId",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "districtName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "schoolType",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "charter",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "minEnrollment",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "maxEnrollment",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "includeInactive",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "Include closed schools"
          },
          {
            "name": "metricYear",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "metricExam",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pssa",
                "keystone"
              ]
            }
          },
          {
            "name": "metricSubject",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "Mathematics"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/schools/{id}": {
      "get": {
        "tags": [
          "entities"
        ],
        "summary": "One school with every year of All Students results and growth",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/schools/{id}/similar": {
      "get": {
        "tags": [
          "entities"
        ],
        "summary": "Nearest schools of the same level and similar size",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 4,
              "maximum": 8
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/schools/map": {
      "get": {
        "tags": [
          "entities"
        ],
        "summary": "One point per located school with the all-grades result for a year, exam, subject, and group",
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "$ref": "#/components/parameters/exam"
          },
          {
            "$ref": "#/components/parameters/subject"
          },
          {
            "name": "group",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "All Students"
            }
          },
          {
            "name": "indicator",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "regular_attendance",
                "grad_rate_4yr",
                "low_income"
              ]
            },
            "description": "Colour by a non-assessment indicator; its value is returned in both proficiency and indicator, growth is null"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/indicators/school/{id}": {
      "get": {
        "tags": [
          "indicators"
        ],
        "summary": "Future Ready PA Index indicators, graduation rate, and enrollment history for one school",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "indicators: series per indicator with statewide value and a percentile among schools; enrollment by year; groups: latest attendance and graduation by student group with the gap vs. All Students"
          }
        }
      }
    },
...
    "/schools/nearby": {
      "get": {
        "tags": [
          "entities"
        ],
        "summary": "Schools closest to a point with their latest Math/ELA (and Keystone) proficiency",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "lng",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 15,
              "maximum": 50
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "year, origin, schools[{id,name,type,city,districtName,lat,lng,km,math,ela,algebra,literature}]"
          }
        }
      }
    }
  }
}